site_construction.cpp

+++ src/modules/site/site_construction.cpp
@@ -48,18 +48,18 @@ void systemUpdateConstructionSiteLocations(flecs::entity entity, Site &site) {
     }
   });
 
-  auto loc = [&site](size_t y, size_t x) -> size_t {
+  auto loc = [&site](uint8_t y, uint8_t x) -> uint8_t {
     return y * site.height + x;
   };
 
-  auto is_valid = [&site](size_t y, size_t x) -> bool {
+  auto is_valid = [&site](uint8_t y, uint8_t x) -> bool {
     if (y > site.height || x > site.width) {
       return false;
     }
     return true;
   };
 
-  auto setConstruction = [&](size_t y, size_t x) {
+  auto setConstruction = [&](uint8_t y, uint8_t x) {
     if (!is_valid(y, x)) {
       return;
     }
@@ -69,8 +69,8 @@ void systemUpdateConstructionSiteLocations(flecs::entity entity, Site &site) {
   };
 
   bool empty_site = true;
-  for (size_t y = 0; y < site.height; y++) {
-    for (size_t x = 0; x < site.width; x++) {
+  for (uint8_t y = 0; y < site.height; y++) {
+    for (uint8_t x = 0; x < site.width; x++) {
       if (locationMap[loc(y, x)] != LocationInfo::TileBuilding) {
         continue;
       }
@@ -87,8 +87,8 @@ void systemUpdateConstructionSiteLocations(flecs::entity entity, Site &site) {
     setConstruction(site.height / 2, site.width / 2);
   }
 
-  for (unsigned int y = 0; y < site.height; y++) {
-    for (unsigned int x = 0; x < site.width; x++) {
+  for (uint8_t y = 0; y < site.height; y++) {
+    for (uint8_t x = 0; x < site.width; x++) {
       if (locationMap[loc(y, x)] != LocationInfo::TileConstruction) {
         continue;
       }
